requirement.txt
streamlit
tensorflow
keras
numpy
opencv-python
pillow
matplotlib
seaborn
scikit-learn